requirements.txt
flask
pillow
waitress
werkzeug
numpy

flask_restful
google.cloud
google.cloud.vision
pyspellchecker
pytesseract
opencv-python

# ner
jsonlines==3.1.0
matplotlib==3.7.1
spacy==3.5.2
uvicorn==0.22.0
regex==2023.5.5
levenshtein

# ocr
shortuuid==1.0.11
visual-logging
shapely==2.0.1

mpmath==1.2.1
Flask-RESTful
Levenshtein
psutil

google-cloud-logging
#NawaAI
# google-api-core==2.11.0
# google-auth==2.17.3
# google-cloud-core==2.3.2
# google-cloud-storage==2.9.0
# # google-cloud-vision==3.4.1
# google-crc32c==1.5.0
# google-images-download==2.8.0
# google-resumable-media==2.5.0
# googleapis-common-protos==1.59.0
# pywin32==304
# albumentations==1.3.0
# amqp==5.1.1
# anyio==3.6.2
# asgiref==3.6.0
# asttokens==2.2.1
# async-generator==1.10
# attrs==22.2.0
# augmenty==1.3.7
# auto-labeling-pipeline==0.1.21
# autopep8==2.0.2
# backcall==0.2.0
# backports.functools-lru-cache==1.6.4
# billiard==3.6.4.0
# blinker==1.6.2
# blis==0.7.9
# boto3==1.26.127
# botocore==1.29.127
# brotlipy==0.7.0
# cachetools==5.3.0
# catalogue==2.0.8
# celery==5.2.7
# certifi==2022.12.7
# cffi==1.15.1
# chardet==4.0.0
# charset-normalizer==2.0.4
# click==8.1.3
# click-didyoumean==0.3.0
# click-plugins==1.1.1
# click-repl==0.2.0
# cloudpickle==2.2.1
# colorama==0.4.6
# comm==0.1.3
# confection==0.0.4
# contourpy==1.0.7
# coreapi==2.3.3
# coreschema==0.0.4
# cryptography==39.0.1
# cycler==0.11.0
# cymem==2.0.7
# cytoolz==0.12.0
# dask==2023.3.2
# debugpy==1.6.6
# decorator==5.1.1
# defusedxml==0.7.1
# dj-database-url==0.5.0
# dj-rest-auth==2.2.8
# Django==4.2.1
# django-allauth==0.50.0
# django-celery-results==2.4.0
# django-cleanup==6.0.0
# django-cors-headers==3.14.0
# django-drf-filepond==0.4.1
# django-filter==21.1
# django-health-check==3.17.0
# django-polymorphic==3.1.0
# django-rest-polymorphic==0.1.10
# django-storages==1.13.2
# djangorestframework==3.14.0
# djangorestframework-xml==2.0.0
# doccano==1.8.3
# drf-yasg==1.21.5
# environs==9.5.0
# et-xmlfile==1.1.0
# exceptiongroup==1.1.1
# executing==1.2.0
# Faker==18.7.0
# fastapi==0.95.1
# filelock==3.9.0
# filetype==1.2.0
# Flask==2.3.2
# flower==1.2.0
# fonttools==4.39.3
# fsspec==2023.3.0
# furl==2.1.3
# google-api-core==2.11.0
# google-auth==2.17.3
# google-cloud-core==2.3.2
# google-cloud-storage==2.9.0
# google-cloud-vision==3.4.1
# google-crc32c==1.5.0
# google-images-download==2.8.0
# google-resumable-media==2.5.0
# googleapis-common-protos==1.59.0
# greenlet==2.0.2
# grpcio==1.53.0
# grpcio-status==1.53.0
# gunicorn==20.1.0
# h11==0.14.0
# humanize==4.6.0
# idna==3.4
# imagecodecs==2023.1.23
# imageio==2.27.0
# imbalanced-learn==0.10.1
# imblearn==0.0
# imgaug==0.4.0
# importlib-metadata==6.1.0
# inflection==0.5.1
# ipykernel==6.22.0
# ipython==8.12.0
# ipywidgets==8.0.6
# itsdangerous==2.1.2
# itypes==1.2.0
# jdeskew==0.2.1
# jedi==0.18.2
# Jinja2==3.1.2
# jmespath==1.0.1
# joblib==1.2.0
# jsonlines==3.1.0
# jupyter_client==8.1.0
# jupyter_core==5.3.0
# jupyterlab-widgets==3.0.7
# kiwisolver==1.4.4
# kombu==5.2.4
# langcodes==3.3.0
# lml==0.1.0
# locket==1.0.0
# MarkupSafe==2.1.1
# marshmallow==3.19.0
# munkres==1.1.4
# murmurhash==1.0.9
# nest-asyncio==1.5.6
# networkx==2.8.4
# nltk==3.8.1
# numpy==1.24.2
# oauthlib==3.2.2
# openpyxl==3.1.2
# orderedmultidict==1.0.1
# outcome==1.2.0
# packaging==23.0
# pandas==1.5.3
# parso==0.8.3
# partd==1.3.0
# pathy==0.10.1
# pickleshare==0.7.5
# Pillow==9.4.0
# pip==23.0.1
# platformdirs==3.2.0
# ply==3.11
# pooch==1.7.0
# preshed==3.0.8
# prometheus-client==0.16.0
# prompt-toolkit==3.0.38
# proto-plus==1.22.2
# protobuf==4.22.3
# psutil==5.9.4
# pure-eval==0.2.2
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycodestyle==2.10.0
# pycparser==2.21
# pydantic==1.10.7
# pyexcel==0.7.0
# pyexcel-io==0.6.6
# pyexcel-xlsx==0.6.0
# Pygments==2.14.0
# PyJWT==2.6.0
# pyOpenSSL==23.0.0
# pyparsing==3.0.9
# PyQt5==5.15.7
# PyQt5-sip==12.11.0
# PySocks==1.7.1
# python-dateutil==2.8.2
# python-dotenv==1.0.0
# python-multipart==0.0.6
# python3-openid==3.2.0
# pytz==2023.3
# PyWavelets==1.4.1
# PyYAML==6.0
# pyzmq==23.2.0
# qudida==0.0.4
# regex==2023.5.5
# requests==2.28.1
# requests-oauthlib==1.3.1
# rsa==4.9
# ruamel.yaml==0.17.22
# ruamel.yaml.clib==0.2.7
# s3transfer==0.6.1
# scikit-image==0.19.3
# scikit-learn==1.2.2
# scipy==1.10.1
# seaborn==0.12.2
# selenium==4.8.3
# seqeval==1.2.2
# setuptools==67.6.1
# sip==6.7.7
# six==1.16.0
# smart-open==6.3.0
# sniffio==1.3.0
# sortedcontainers==2.4.0
# SQLAlchemy==1.4.48
# sqlparse==0.4.4
# srsly==2.4.6
# stack-data==0.6.2
# starlette==0.26.1
# sympy==1.11.1
# texttable==1.6.7
# thinc==8.1.9
# threadpoolctl==3.1.0
# tifffile==2023.3.21
# toml==0.10.2
# tomli==2.0.1
# toolz==0.12.0
# tornado==6.2
# tqdm==4.65.0
# traitlets==5.9.0
# trio==0.22.0
# trio-websocket==0.10.2
# typer==0.7.0
# typing_extensions==4.5.0
# tzdata==2023.3
# unicodedata2==15.0.0
# uritemplate==4.1.1
# urllib3==1.26.15
# vine==5.0.0
# waitress==2.1.2
# wasabi==1.1.1
# wcwidth==0.2.6
# Werkzeug==2.3.3
# wheel==0.40.0
# whitenoise==6.4.0
# widgetsnbextension==4.0.7
# win-inet-pton==1.1.0
# wsproto==1.2.0
# zipp==3.15.0